Complete logic extraction from diary
Extracts functions from diary.js into diary-food-add.js,
diary-food-remove.js, diary-food-search.js, and diary-food-table.js.
Functions are logically broken into each of the remainging files in
the diary folder. Necessary exports and changes to function
invocation were performed in order to maintain proper functionality.
All page behavior remains appropriate to specifications.
