+++ lib/diary/diary-food-add.js
@@ -61,6 +61,23 @@ const calories = require('./calories')
     calories.updateConsumedCalories()
   }
 
+  $(document).ready( () => {
+
+      $('#Breakfast').on('click', (event) => {
+        addMeal()
+      })
+      $('#Lunch').on('click', (event) => {
+        addMeal()
+      })
+      $('#Dinner').on('click', (event) => {
+        addMeal()
+      })
+      $('#Snacks').on('click', (event) => {
+        addMeal()
+      })
+    $('form').on('submit', (event) => event.preventDefault());
+  });
+
   module.exports = {addMeal: addMeal,
                     grabFood: grabFood,
                     addFoodstoMealTable: addFoodstoMealTable,
