+++ lib/diary/diary-food-remove.js
@@ -27,4 +27,9 @@ function removeFoodFromMeal(){
   })
 }
 
-module.exports = {removeFoodFromMeal: removeFoodFromMeal}
+$(document).ready( () => {
+  $('body').on('click','.remove', (event) => {
+    removeFoodFromMeal()
+  })
+  $('form').on('submit', (event) => event.preventDefault());
+});
