+++ lib/diary/diary.js
@@ -3,11 +3,11 @@ const LOCAL = 'http://localhost:3000/api/v1/'
 const diaryFoodTable = require('./diary-food-table')
 const calories = require('./calories')
 const foodAdder = require('./diary-food-add')
-const foodRemover = require('./diary-food-remove')
+require('./diary-food-remove')
+require('./diary-food-search')
 $ = require('jquery')
 
 let counter = 0
-
 
 const Diary = class Diary {
 
@@ -126,47 +126,13 @@ function populateMealsTables(data){
   }
 }
 
-function searchFoods(){
-  let searchName = $('#food-filter').val()
-  $('#diary-foods-body').html('')
-  return $.ajax({
-    url: `${API}search?searchName=${searchName}`,
-    method: 'GET',
-  })
-  .then((data) => {
-    diaryFoodTable.makeFoodTable(data)
-  })
-  .fail( (error) => {
-    console.error(error)
-  })
-}
-
 $(document).ready(function(){
   const diary = new Diary
-
-  $('#food-filter').on('input',function(){
-    searchFoods()
-  })
 
   diaryFoodTable.getFoods()
 
   diary.getDiary()
 
-  $('#Breakfast').on('click', (event) => {
-    foodAdder.addMeal()
-  })
-  $('#Lunch').on('click', (event) => {
-    foodAdder.addMeal()
-  })
-  $('#Dinner').on('click', (event) => {
-    foodAdder.addMeal()
-  })
-  $('#Snacks').on('click', (event) => {
-    foodAdder.addMeal()
-  })
-  $('body').on('click','.remove', (event) => {
-    foodRemover.removeFoodFromMeal()
-  })
   $('body').on('click', '#prev-page-btn', (event) => {
     counter = counter - 1
     diary.getPrevious()
